fix(uml): fix code reported by CodeQL
Fixes:

- https://github.com/pdfclown/jada-examples/security/code-scanning/1

- https://github.com/pdfclown/jada-examples/security/code-scanning/2

- https://github.com/pdfclown/jada-examples/security/code-scanning/3

- https://github.com/pdfclown/jada-examples/security/code-scanning/4

- https://github.com/pdfclown/jada-examples/security/code-scanning/5

+++ jada-examples-uml/src/main/java/org/pdfclown/jada/examples/uml/protoevo/biology/Cell.java
@@ -107,7 +107,7 @@ public abstract class Cell extends Particle implements Serializable {
         children.add(child);
       } catch (MiscarriageException ignored) {
       }
-      angle += 2 * Math.PI / nChildren;
+      angle += (float) (2 * Math.PI / nChildren);
     }
 
     for (int j = 0; j < 8; j++)
